VAT - changed from integer to numeric in booking_invoice_item

+++ database/migrations/2016_05_26_193853_add_fields_to_booking_invoice_items.php
@@ -19,8 +19,8 @@ class AddFieldsToBookingInvoiceItems extends Migration
             $table->date('booking_date')->after('quantity');
             $table->string('booking_time_interval', 50)->after('booking_date');
             $table->float('price')->after('booking_time_interval');
-            $table->tinyInteger('vat')->after('price');
-            $table->tinyInteger('discount')->after('vat');
+            $table->float('vat')->after('price');
+            $table->float('discount')->after('vat');
             $table->float('total_price')->after('discount');
         });
     }
